+++ app/controllers/users/messages_controller.rb
@@ -1,6 +1,7 @@
 # Allowing users to send user-to-user messages
 class Users::MessagesController < UserController
   before_action :set_recipient,
+                :check_messaging_enabled,
                 :check_recipient_accepts_messages,
                 :check_can_send_messages,
                 :check_logged_in,
@@ -33,6 +34,21 @@ class Users::MessagesController < UserController
 
   def set_recipient
     @recipient_user = User.find_by!(url_name: params[:url_name])
+  end
+
+  def signed_recipient
+    return unless params[:sgid]
+
+    GlobalID::Locator.locate_signed(
+      params[:sgid], for: 'user_to_user_messaging'
+    )
+  end
+
+  def check_messaging_enabled
+    return if feature_enabled?(:user_to_user_messaging)
+    return if @recipient_user == signed_recipient
+
+    render template: 'users/messages/disabled'
   end
 
   def check_recipient_accepts_messages