+++ lib/configuration.rb
@@ -49,6 +49,7 @@ module AlaveteliConfiguration
       ENABLE_PRO_PRICING: false,
       ENABLE_PRO_SELF_SERVE: false,
       ENABLE_TWO_FACTOR_AUTH: false,
+      ENABLE_USER_TO_USER_MESSAGING: true,
       ENABLE_WIDGETS: false,
       EXCEPTION_NOTIFICATIONS_FROM: 'errors@localhost',
       EXCEPTION_NOTIFICATIONS_TO: 'user-support@localhost',