+++ doc/CHANGES.md
@@ -2,6 +2,7 @@ j develop
 
 ## Highlighted Features
 
+* Add user-to-user messaging configuration (Graeme Porteous)
 * Add `rel=nofollow` attributes to links in outgoing messages (Graeme Porteous)
 * Add task to purge profile content for limited user profiles after six months
   (Graeme Porteous)